Refactor moved config, infrastructure and main.go in root folder
